Only get the most recent reference data for each statement
The results from the query are sorted by date, so we can use `qsv dedup`
to only get the most recent for each psid.

This means if someone's name changes upstream, we can add a new
reference with this `named as`, and that will be the one that gets
returned in future.